requirements.txt
albumentations==0.4.3
clip==0.2.0
einops==0.3.0
fire==0.5.0
imwatermark==0.0.2
invisible_watermark==0.1.5
ipython==8.12.0
ipywidgets==8.0.6
kornia==0.6.0
matplotlib==3.7.1
natsort==8.3.1
numpy==1.24.2
omegaconf==2.1.1
openai_clip==1.0.1
opencv_python
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
protobuf
pudb==2019.2
pyarrow==11.0.0
pytorch_lightning==1.4.2
PyYAML==6.0
scann==1.2.9
scipy==1.10.1
setuptools==65.6.3
taming_transformers_rom1504==0.0.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
transformers==4.19.2
torchmetrics==0.6.0
test-tube
# removebg
rembg[gpu]
pygsheets<=1.5
oauth2client
